refactor: update images and refactor translation buttons in hippotherapy and who-we-are sections
