+++ src/pages/admin/programs/components/program-list-item/ProgramListItem.scss
@@ -127,14 +127,6 @@
                 height: 1.5rem;
                 border: none;
             }
-
-            .translate-btn {
-                background: url('~@/assets/icons/translate.svg');
-
-                &:active {
-                    opacity: 0.8;
-                }
-            }
         }
     }
