+++ src/pages/admin/who-we-are/components/modals/strategies/description/translate-multiple-descriptions-strategy.ts
@@ -5,13 +5,13 @@ import {
 } from '../../forms/translate-multiple-descriptions-form/TranslateWhoWeAreMultipleDescriptionsForm';
 import { WhoWeAreModalStrategy } from '../who-we-are-modal-strategy';
 import { SectionType } from '@/types/common/about-us';
-import SupportVeterans from '@/assets/images/public/about-us-page/support-veterans.jpg';
-import SupportVolunteers from '@/assets/images/public/about-us-page/support-volunteers.jpg';
-import SupportChildren from '@/assets/images/public/about-us-page/support-children.jpg';
-import ManAndHorse from '@/assets/images/public/about-us-page/man-horse.jpg';
-import GirlAndHorse from '@/assets/images/public/about-us-page/girl-horse.jpg';
-import OldManAndHorse from '@/assets/images/public/about-us-page/old-man-horse.jpg';
-import WomanAndHorse from '@/assets/images/public/about-us-page/woman-horse.jpg';
+import SupportVeterans from '@/assets/images/support-veterans.webp';
+import SupportVolunteers from '@/assets/images/support-volunteers.webp';
+import SupportChildren from '@/assets/images/support-children.webp';
+import ManAndHorse from '@/assets/images/man-horse.webp';
+import GirlAndHorse from '@/assets/images/girl-horse.webp';
+import OldManAndHorse from '@/assets/images/old-man-horse.webp';
+import WomanAndHorse from '@/assets/images/woman-horse.webp';
 
 export const translateMultipleDescriptionsStrategy: WhoWeAreModalStrategy<TranslateWhoWeAreMultipleDescriptionsFormValues> =
     {
