+++ src/pages/public/hippotherapy-page/HippotherapyPage.tsx
@@ -6,8 +6,8 @@ import { HippotherapyIntro } from './intro/HippotherapyIntro';
 import { LoadableContent } from '@/components/common/loadable-content/LoadableContent';
 import { TextCard } from './text-card/TextCard';
 import { QuoteSection } from './quote/QuoteSection';
-import quote1DefaultImg from '@/assets/images/public/hippotherapy/quote_1.jpg';
-import quote2DefaultImg from '@/assets/images/public/hippotherapy/quote_2.jpg';
+import quote1DefaultImg from '@/assets/images/man-holding-horse-cheek-close.webp';
+import quote2DefaultImg from '@/assets/images/man-facing-horse-forehead.webp';
 import { SloganSection } from './slogan/SloganSection';
 import { HippoventionCenter } from './hippovention-center/HippoventionCenter';
